+++ ocfl-validate.py
@@ -1,15 +1,16 @@
 #!/usr/bin/env python
 """Validate an OCFL Object."""
 import argparse
+import logging
 import ocfl
 import sys
 
 parser = argparse.ArgumentParser(
-    description='Validate one or more OCFL objects. By default shows any '
-    'errors or warnings, and final validation status. Use -q to show only errors, '
-    '-Q to show only validation status.')
-parser.add_argument('objdir', type=str, nargs='*',
-                    help='OCFL object path(s) of objects to validate')
+    description='Validate one or more OCFL objects or storage roots. By default '
+    'shows any errors or warnings, and final validation status. Use -q to show '
+    'only errors, -Q to show only validation status.')
+parser.add_argument('path', type=str, nargs='*',
+                    help='OCFL object or storage root path(s) to validate')
 parser.add_argument('--quiet', '-q', action='store_true',
                     help="be quiet, do not show warnings")
 parser.add_argument('--very-quiet', '-Q', action='store_true',
@@ -23,15 +24,39 @@ ocfl.add_shared_args(parser)
 args = parser.parse_args()
 ocfl.check_shared_args(args)
 
-if len(args.objdir) == 0:
-    print("No OCFL bject paths specified, nothing to do! (Use -h for help)")
-ocfl = ocfl.Object(lax_digests=args.lax_digests)
-num_bad = 0
-for objdir in args.objdir:
-    if not ocfl.validate(objdir,
-                         show_warnings=not args.quiet and not args.very_quiet,
-                         show_errors=not args.very_quiet,
-                         check_digests=not args.no_check_digests):
-        num_bad += 1
-if num_bad > 0:
+logging.basicConfig(level=logging.INFO if args.verbose else logging.WARN)
+log = logging.getLogger(name="ocfl-validate")
+log.setLevel(level=logging.INFO if args.verbose else logging.WARN)
+
+if len(args.path) == 0:
+    print("No OCFL paths specified, nothing to do! (Use -h for help)")
+
+num = 0
+num_good = 0
+num_paths = len(args.path)
+for path in args.path:
+    num += 1
+    path_type = ocfl.find_path_type(path)
+    if path_type == 'object':
+        log.info("Validating OCFL Object at " + path)
+        obj = ocfl.Object(lax_digests=args.lax_digests)
+        if obj.validate(path,
+                        show_warnings=not args.quiet and not args.very_quiet,
+                        show_errors=not args.very_quiet,
+                        check_digests=not args.no_check_digests):
+            num_good += 1
+    elif path_type == 'root':
+
+        log.info("Validating OCFL Storage Root at " + path)
+        store = ocfl.Store(root=path,
+                           lax_digests=args.lax_digests)
+        if store.validate(show_warnings=not args.quiet and not args.very_quiet,
+                          show_errors=not args.very_quiet,
+                          check_digests=not args.no_check_digests):
+            num_good += 1
+    else:
+        log.error("Bad path %s (%s)" % (path, path_type))
+    if num_paths > 1:
+        log.info(" [%d / %d paths validated, %d / %d VALID]\n" % (num, num_paths, num_good, num))
+if num_good != num:
     sys.exit(1)