+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## ... v0.0.6
 
-  * ...
+  * Handle case of no versions https://ocfl.io/draft/spec/#E008
 
 ## 2020-05-05 v0.0.5